pyproject.toml
[project]
name = "marchmadnesspredictorrlvr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "verifiers>=0.0.0",
    "google-api-python-client>=2.0.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.0.0",
]

[tool.uv.workspace]
members = [
    "test",
]
